Reorganized test and style files
