+++ .vscode/symbols.json
@@ -1,0 +1,1 @@
+{"symbols":{},"files":{}}
